+++ CHANGELOG.md
@@ -1,8 +1,34 @@
-# [2.2.0](https://github.com/Bugs5382/node-hl7-client/compare/v2.1.0...v2.2.0) (2024-06-04)
+# [2.3.0-beta.1](https://github.com/Bugs5382/node-hl7-client/compare/v2.2.0...v2.3.0-beta.1) (2024-06-24)
 
 
 ### Bug Fixes
 
+* clear connection timer ([9d9f296](https://github.com/Bugs5382/node-hl7-client/commit/9d9f2965906417b4027640950e63d8d9ca7f5314)), closes [#100](https://github.com/Bugs5382/node-hl7-client/issues/100)
+* connections and threads failing to close ([#102](https://github.com/Bugs5382/node-hl7-client/issues/102)) ([013d95e](https://github.com/Bugs5382/node-hl7-client/commit/013d95ec0a57f7150d48e17e0248b793380219e0))
+* lint issue ([2a35e19](https://github.com/Bugs5382/node-hl7-client/commit/2a35e1935ab3071bfa487478f2ae5184add1f200))
+* lint issue ([80d7667](https://github.com/Bugs5382/node-hl7-client/commit/80d7667ba32a706783b44ee8644cf23d7eecf0ec))
+* race condition ([4b9e6cf](https://github.com/Bugs5382/node-hl7-client/commit/4b9e6cf880973c3e8921f696c6f467f1cc595c63)), closes [#100](https://github.com/Bugs5382/node-hl7-client/issues/100)
+* race condition ([84f0a89](https://github.com/Bugs5382/node-hl7-client/commit/84f0a898670315fec4d7660d89efcdcfab180d7d)), closes [#100](https://github.com/Bugs5382/node-hl7-client/issues/100)
+* race condition ([#106](https://github.com/Bugs5382/node-hl7-client/issues/106)) ([1a8bf3e](https://github.com/Bugs5382/node-hl7-client/commit/1a8bf3e267bfcff2be1b11ab98a255fc7480ba8a))
+
+
+### Features
+
+* minor code cleanups ([#105](https://github.com/Bugs5382/node-hl7-client/issues/105)) ([da37ee1](https://github.com/Bugs5382/node-hl7-client/commit/da37ee101b959ff7a131775aa7b2c2a1b971ae1c))
+* typedoc working! ([6b884ff](https://github.com/Bugs5382/node-hl7-client/commit/6b884ff6c5a6e43bbb52c4de37a8afec229c599e)), closes [#37](https://github.com/Bugs5382/node-hl7-client/issues/37)
+
+# [2.2.0-beta.1](https://github.com/Bugs5382/node-hl7-client/compare/v2.1.1-beta.3...v2.2.0-beta.1) (2024-06-24)
+
+# [2.2.0](https://github.com/Bugs5382/node-hl7-client/compare/v2.1.0...v2.2.0) (2024-06-04)
+
+### Bug Fixes
+
+
+* lint issue ([2a35e19](https://github.com/Bugs5382/node-hl7-client/commit/2a35e1935ab3071bfa487478f2ae5184add1f200))
+* lint issue ([80d7667](https://github.com/Bugs5382/node-hl7-client/commit/80d7667ba32a706783b44ee8644cf23d7eecf0ec))
+* race condition ([4b9e6cf](https://github.com/Bugs5382/node-hl7-client/commit/4b9e6cf880973c3e8921f696c6f467f1cc595c63)), closes [#100](https://github.com/Bugs5382/node-hl7-client/issues/100)
+* race condition ([84f0a89](https://github.com/Bugs5382/node-hl7-client/commit/84f0a898670315fec4d7660d89efcdcfab180d7d)), closes [#100](https://github.com/Bugs5382/node-hl7-client/issues/100)
+* race condition ([#106](https://github.com/Bugs5382/node-hl7-client/issues/106)) ([1a8bf3e](https://github.com/Bugs5382/node-hl7-client/commit/1a8bf3e267bfcff2be1b11ab98a255fc7480ba8a))
 * 2.4 spec MSH9.3 ([7af0970](https://github.com/Bugs5382/node-hl7-client/commit/7af097037e201b7fb63394ea757c2edf688a965c))
 * parsing error ([adb0911](https://github.com/Bugs5382/node-hl7-client/commit/adb09115964b7857e33c6b11b1db84bc92beeefe)), closes [#96](https://github.com/Bugs5382/node-hl7-client/issues/96)
 * parsing error ([#97](https://github.com/Bugs5382/node-hl7-client/issues/97)) ([747ef5d](https://github.com/Bugs5382/node-hl7-client/commit/747ef5d33fe72ba5f3c3434563fa1ef669652364))
@@ -11,6 +37,17 @@
 
 ### Features
 
+
+* minor code cleanups ([#105](https://github.com/Bugs5382/node-hl7-client/issues/105)) ([da37ee1](https://github.com/Bugs5382/node-hl7-client/commit/da37ee101b959ff7a131775aa7b2c2a1b971ae1c))
+* typedoc working! ([6b884ff](https://github.com/Bugs5382/node-hl7-client/commit/6b884ff6c5a6e43bbb52c4de37a8afec229c599e)), closes [#37](https://github.com/Bugs5382/node-hl7-client/issues/37)
+
+## [2.1.1-beta.3](https://github.com/Bugs5382/node-hl7-client/compare/v2.1.1-beta.2...v2.1.1-beta.3) (2024-06-22)
+
+
+### Bug Fixes
+
+* clear connection timer ([9d9f296](https://github.com/Bugs5382/node-hl7-client/commit/9d9f2965906417b4027640950e63d8d9ca7f5314)), closes [#100](https://github.com/Bugs5382/node-hl7-client/issues/100)
+* connections and threads failing to close ([#102](https://github.com/Bugs5382/node-hl7-client/issues/102)) ([013d95e](https://github.com/Bugs5382/node-hl7-client/commit/013d95ec0a57f7150d48e17e0248b793380219e0))
 * merge in fixes ([#98](https://github.com/Bugs5382/node-hl7-client/issues/98)) ([97bbb58](https://github.com/Bugs5382/node-hl7-client/commit/97bbb589a4bcf50cb9f9228cba1f82f88e9ea259))
 
 ## [2.1.1-beta.2](https://github.com/Bugs5382/node-hl7-client/compare/v2.1.1-beta.1...v2.1.1-beta.2) (2024-06-04)